webadmin: fixed initial focus in edit pool dialog
The name textbox is the default component to get the initial focus in vm
dialogs, but in the edit pool dialog the name textbox is disabled,
thus no component gets focus when the dialog is opened.
To fix this, I changed the component that gets the initial focus in the
edit pool dialog to the description textbox.

Change-Id: I55445daaf90d0c394f6da28cc3d39495b1e3c115
Signed-off-by: Arik Hadas <ahadas@redhat.com>