core: GetAllFromExportDomainQueryParameters.getAll
Removed the GetAll parameter from the
GetAllFromExportDomainQueryParameters class. This parameter was ALWAYS
set to true, with no way for the user to set it to false.
Thus, the parameter itself is useless (it's effectively hardcoded to
true), and all the code that handles GetAll=false is dead code that
cannot be reached.

This patch removes this useless code, as it obviously has no use in the
system. A neat side-effect of removing this code is saving a couple of
VmDao.getAll() and VmTemplateDao.getAll() database calls.

Change-Id: I3cb36d7a5b1a9e309a7553fc2ef696b1cc1cae5a
Signed-off-by: Allon Mureinik <amureini@redhat.com>