core: ImportVm test cleanup: Don't mock DAOs
Cleaned up the TestHelperImprotVmCommand to stop mocking DAOs and
instead override methods that represent a higher business logic, in
order to make the test more readable and understandable.

In order to achieve this refactoring, a few blocks of
ImportVmCommand.canDoAction() were extracted to methods, which could be
overridden.

Note that this is an interim patch, performed in order to make the real
goal, removing TestHelperImportVmCommand altogether.

Change-Id: I19d711cd8df8c0f8eccfe6b7838c16605829b2f6
Signed-off-by: Allon Mureinik <amureini@redhat.com>