core:encrypt vds_static pm_password field on ...
core:encrypt vds_static pm_password field on upgrade to 3.1

This patch sets the correct configuration values in vdc_options for the
03_00_0420...sql upgrade script.
Since those values are set only in the installer post upgrade step, we
must update them in the DB pre-upgrade step in order to encrypt passwords
successfully.

This patch also removes one of those entries from engine-config tool
because the entry should be read-only.

Change-Id: I5d2cb10f205a90ae0752498a9a3d3f0eda0ec9fb
Signed-off-by: Eli Mesika <emesika@redhat.com>