webadmin: DataProvider.java cleanup (2/6)
DataProvider is a class which is mostly obsolete,
as it contains synchronous calls to the engine,
which are not relevant anymore under the GWT platform/infrastructure.

The goal is to eliminate this class all-together.

In this patch:
1) Removing methods that used old cache mechanism for
config values, and replacing them with the new mechanism
in AsyncDataProvider (only 2 methods)
2) Removing DataProvider config values mechanism

Change-Id: I8eb69268ef416c1738d92129e8c10f3197967a53
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>
Bug-Url: https://bugzilla.redhat.com/881747