webadmin: DataProvider.java cleanup (4/6)
DataProvider is a class which is mostly obsolete,
as it contains synchronous calls to the engine,
which are not relevant anymore under the GWT platform/infrastructure.

The goal is to eliminate this class all-together.

In this patch:
removing sync getTemplateById and getClusterById, which were
used by Edit Pool;
PoolListModel.Edit used these methods while filling out the
dialog incase the pool is empty. We cannot edit empty pool, hence
removed the code dealing with it.

Change-Id: Ie7dbd26e10a5d0680a74c5a8dc1ee5cc9461c309
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>
Bug-Url: https://bugzilla.redhat.com/881747