webadmin: DataProvider.java cleanup (5/6)
DataProvider is a class which is mostly obsolete,
as it contains synchronous calls to the engine,
which are not relevant anymore under the GWT platform/infrastructure.

The goal is to eliminate this class all-together.

In this patch:
Removing sync call for IsVmNameUnique, and replacing it with
an async call in AsyncDataProvider

Change-Id: If11911f94e02a5c115a4c4407b918580704bd9eb
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>
Bug-Url: https://bugzilla.redhat.com/881747