webadmin: DataProvider.java cleanup (6/6)
DataProvider is a class which is mostly obsolete,
as it contains synchronous calls to the engine,
which are not relevant anymore under the GWT platform/infrastructure.

The goal is to eliminate this class all-together.

In this patch:
1) Removing sync getUsers and not replacing it;
The original flow was to get all users, and filter
out users that are already included.
This filtering needs further thinking,
beacuse the admins may be confused by it (e.g. empty list)

2) Deleting DataProvider.java since it's now empty

Change-Id: I31f6af5b3af32a55a623ce701c2a6c384bef6ddc
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>
Bug-Url: https://bugzilla.redhat.com/881747