core: Fix IndexOutOfBoundsException in plug FC disk
Attach/Detach and plug disk throws an indexOutOfBoundsException.
The reason for that is that the FC LUN which the disk is related to does not contains any
connections in it, as opposed to ISCSI lun which must have connections
related to it.

The proposed fix is to use FCP storage type if the LUN does not contain
any connections in it.

Change-Id: I6414445b9a74be299205ff7fc9a21d1388a29687
Bug-Url: https://bugzilla.redhat.com/882825
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>