tools: DB connections improvement at notifier
A performance issue (and maybe not the only one) exists at
NotificationService:
There is a loop over the result-set representing the
event subscrirbers. As long as the result-set object is active
the connection is held. in each iteration there is an attempt to
send mail - this operation may take time.
The patch seperates the DB  extraction of subscribers from the logic
that is done with each one of them.
This patch also adds a close to a connection at the method
updateAuditLogEventProcessed

Bug-Id: https://bugzilla.redhat.com/show_bug.cgi?id=882847
Change-Id: I54fb540ee52255f017933b000b9b509c2e056122
Signed-off-by: Yair Zaslasky <yzaslavs@redhat.com>