perform connectStorageServer to hosts after during reconstruct/recovery(#880180)
When performing reconstruct/reinitialize DC, a copy of the pool metadata
is being done to a data domain, if hosts doesn't have the needed
'connections' to the new master storage server they will move eventually to non-operational state.

This patch adds execution of connectStorageServer to all hosts other
then the host that executed the reconstruct/recovery (as this host
managed to perform reconstruct successfully, it means that he has the
needed 'connections' to the storage server).

Change-Id: I3733b732b37f3a687b24f90bfdb1799757434d75
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=880180
Signed-off-by: Liron Aravot <laravot@redhat.com>