webadmin: Setup Network- does not show the correct boot protocol.
The fix in commit 30f20efd648a1329c996f5c12f31741038bdc0be caused another
bug.
The attach method is used not only for manually attaching but also in the
init dialog process.
That's why the boot protocol was always set to the default values ("dhcp"
for management and "none" for the others) and never showed the real values.

The boot protocol should be reseted to the default values just if the
current boot protocol is null.

Change-Id: I40fe51b46f94bdfa27270c9e5dcd94074b881c03
Bug-Url:https://bugzilla.redhat.com/875724
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>