gluster: Introduced locking on gluster clusters
GlusterManager can potentially interfere with BLL commands that
execute a gluster command and then update DB with appropriate
changes. Hence it is required that all such gluster BLL commands
acquire a lock before performing their logic, and release it in
the end.

To enable this, added the @LockIdNameAttribute annotation on all
gluster commands that require the locking, and overrode the method
getExclusiveLocks() in GlusterCommandBase and GlusterVolumeCommandBase
to make sure that the lock is acquired on the id of the gluster cluster
on which the action is being performed.

Also modified the GlusterManager to acquire and release locks at
appropriate places when syncing information with GlusterFS.

Introduced a new LockingGroup GLUSTER for all gluster related locking.

Change-Id: I57b3bd0b9e384ff130c4083147ac228a81185a2d
Signed-off-by: Shireesh Anjal <sanjal@redhat.com>