restapi: fix usb mapping on update vm call
Change the way usb policy is set on update vm/template request which is
received through rest api. The full mapping is described in the bug-url.

Specifically, the new mapping states that if no usb setting is defined
in the update request, the previous usb policy will remain as is
(instead of being set to DISABLED as it was before), thus it solves
bug #875881.

UsbResourceUtils class methods moved to VmMapper class, and it was
removed on the way (the same was done for UsbResourceUtilsTest class,
its methods moved to VmMapperTest class).

Change-Id: If74bf6ede2fc0f932db8dfcd7951b6bcf90b7744
Bug-Url: https://bugzilla.redhat.com/875881
Signed-off-by: Arik Hadas <ahadas@redhat.com>