webadmin: "No message" is displayed in Japanese environment
New/Edit VM dialog -> fill a japanese chars into the description
field. An error dialog with message "No message" will be displayed.

The problem was, that the ErrorTranslator expected, that when a message
is in form: "$AAA b" is a variable $AAA with value "b". Than it expected
that the ${AAA} in other messages can be replaced by "b". The problem
was, that the decision, if the specific message is a variable declaration
or a message was decided in a following way:
- if the string starts with "$" character, than it is a variable declaration
- it is a message otherwise

The problem with this is, that when a message looks like "${AAA} b"
than it is considered to be a variable declaration, even it is a message
with reference to the declaration.

Fixed by enriching this decision if somethin is a variable declaratition to
the following:
- if the string starts with "$" but it is does not start with something like
"${some string}" it is a variable
- it is a message otherwise

Change-Id: I04b32be4cf4c05873bcf9643dd5b9acb3695a54a
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>