userportal: allow connection to guest with non resp agent
Change the way we treat non responding agent error received from VDSM
when trying to connect to guest using spice. Instead of showing an error
dialog and canceling the spice connection, we now present a confirmation
dialog (saying that the user will have reduced capabilities due to the
fact the agent is non-responsibe) - if the user choose to cancel, the
spice console won't opened, and if the user choose to continue, spice
console will be opened without SSO.

FrontendEventsHandlerImpl was changed such that it won't show error
dialog in case the returned error is 'nonresp' (non responding guest
agent) from logon command.

In addition, the confirmation dialog described above for the spice console
connection operation was added.

Change-Id: I4eaf009f5ff99946dbb421459ccdec9c536aea58
Bug-Url: https://bugzilla.redhat.com/871083
Signed-off-by: Arik Hadas <ahadas@redhat.com>