engine: Introducing a queue for failovers event
The following patch will introduce a queue for failovers event.
The Idea is to use the follwoing queue in order to ensure a correct order of
the following operations: reconstruct, domain recovery, domain monitoring, etc...
The appropriate event should be submitted to queue and will be executed accorrding to
internal logic of the queue.
Because of the execution of events is done in a separate thread no deadlock should occurred, also
no need for excplicit additional locking
This is the first patch it will introduce a queue and all stuff should work as before, but
using of queue with small change will allow: to reject unneeded events when some othe event is running,
ensure order, ensure appropriate release of quartz threads, fix a races
The following patch also introduce a using of queue, from different places, including RecoveryStoragePoolCommand

Change-Id: Ic1224feacdcdaaaf0b59d26105805ba7ef2a2fff
Bug-Url: https://bugzilla.redhat.com/882837
Signed-off-by: Michael Kublin <mkublin@redhat.com>