webadmin: fix layout of multiple lines conf msgs
Fix the layout of messages in some confirmations dialogs that should be
splitted over multiple lines (such as in the "import vm" confirmation
dialog).

The problem was that the widget which is used to contain those messages
is Label, thus the '\n' characters in the message are ignored, and the
message is presented in a single long line.

This patch solves the problem by replacing the '\n' characters in the
message with "<br>" tags and setting the message as HTML

Change-Id: I1576b7576cf93ba9e525f20261d423fd4c0721b0
Signed-off-by: Arik Hadas <ahadas@redhat.com>