webadmin: Adding confirmation for replicate gluster volumes
While creating a Replicate or Distributed Replicate volume,
if more than one brick is added from the same server,
a confirmation popup will be shown. The same warning will
be shown while adding bricks after creating volume.

In the Add Bricks dialog, a NOTE will be shown in the
bottom explaining the Replicate/Stripe groups will be
created based on the order of the bricks.

Change-Id: Id783c1404eb79b32981bbf137d4228675f3e960f
Bug-Url: https://bugzilla.redhat.com/871261
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>