engine: Adding  VmGuestInfoInterface
As part of the Reporting Internal Vnic Information feature,
this patch adds a new table, Dao and entity called
vm_guest_agent_interface, VmGuestAgentInterfaceDao and
VmGuestAgentInterface respectively.

Each row in this table is related to a specific VM, and will
be presented through REST API under that VM. Also, it will
be associated with Vm Interfaces according to Mac address.

Change-Id: I0d25018f94c2b4d29c01be85a222460335603569
Signed-off-by: Moti Asayag <masayag@redhat.com>