engine: if connect storage pool fails on version mismatch, do reconstruct master
The following patch will allow to perform reconstruct from InitVdsOnUp, in case that we
can not connect to pool because of wrong master version or missing master domain.
Using of event queue will protect from parallel reconstructs and from other races.
The flow is a following:
1. Try connect to pool
2. If failed on wrong master domain or its version - do reconstruct
3. If success move host to UP

Please note: I remove an update of vms at spm after reconstruct, because of it will be removed at the
future and that code is actually failed a one host scenario.

Change-Id: Ia0d11c068399664daf85a6b03e67d25cd382707e
Bug-Url: https://bugzilla.redhat.com/882837
Signed-off-by: Michael Kublin <mkublin@redhat.com>