core: VdsSelector changes for CPU Thread Handling (2/4)
This patch series adds support for CPU Thread Handling in which host
CPU threads can be considered as cores on a per-cluster basis for the
purpose of VM startup/shutdown/migration.

Patch 2: update host selection algorithm to take advantage of cpu
threads.

VdsSelector now uses an "effective" core count.  This is a direct
plug-in replacement for using the physical CPU count, and is only in
effect when the host CPU has SMT enabled, and the cluster is set to
count CPU threads as cores.

Change-Id: I7171f066d646378664e5832ba00fa9f436509b97
Signed-off-by: Greg Padgett <gpadgett@redhat.com>