userportal: NPE in OsTypeImage
The OsTypeImage creates the name of the default OS type
like: "otherOs" + nameUniquePart + IMAGE, where the IMAGE is "Image",
the nameUniquePart is "Large" or "Small".

Than takes this name and looks up the specific image from the
ApplicationResourcesWithLookup. The probelm was, that the otherOsSmallImage
was written with capital "O", which caused periodic NPE thrown on
FE.

Fixed by expecting the "OtherOs..." on OsTypeImage and adjusted also
other names in ApplicationResourcesWithLookup.

Change-Id: I2690cda0142f5c7849b3ad4bcc251906658244f1
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>