core: Add event for MAC pool initialization error
This patch performs two issues:
a. Fixing inconsistency in handling MacPoolManager.freeMac
(in some cases - the exception is caught and ignored, in some cases -
not - this is behavior is incorrect).
b. Introducing event to mark failure in mac pool initialization (similar to the event log
of mac address already in use)

Bug-Url: https://bugzilla.redhat.com/885441
Change-Id: Ibaa3d9eb23016b4866e5cfa39f919378e09fdf2d
signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>