webadmin: allow changing name of VM attached to pool
This is a quick solution for a requirement to be able to manage the
names of VMs which are attached to pools.

We let the user open edit dialog for VM which is attached to pool, and
in this dialog the only attribute the user can change is the VM's
name (all other widgets are disabled).

Change-Id: I139ca66f945657b4d2323d6669e595c77a8fad09
Bug-Url: https://bugzilla.redhat.com/867308
Signed-off-by: Arik Hadas <ahadas@redhat.com>