webadmin: allow a single user lookup at a time
In search user (add user/permission),
a query ('GO') may take a while, so we will allow only
a single search at a time, to avoid multiple concurrent
queries, If a query takes a long time a user may
intuitively press a lot of times on the go button/enter key.

Because the number of connections is limited, we don't
want to waste them with redundant calls.

The 'GO' button will be disabled according to a field in the model
(searchInProgress)

Change-Id: I4a44304d3ffc9fd8e76afe3f8bd443553e64f6f6
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>