core: Fix Remove Storage error message(#863097)
setActionMessageParameters() was not set for RemoveStorageDomainCommand,
nor for base class StorageDomainCommandBase().
When the lock could not be acquired in CommandBase.acquireLockInternal()
both VAR__TYPE__STORAGE__DOMAIN and VAR__ACTION__REMOVE were not added
to the VdcBllMessages, resulting in ${action} and ${type} appearing in
the error message.

This patch implements setActionMessageParameters() in
StorageDomainCommandBase and in RemoveStorageDomainCommand, and removes
the setting of VAR__TYPE__STORAGE__DOMAIN and VAR__ACTION__REMOVE in
other flows, as it is not needed there anymore.

Adjusted and added tests, also added assert messages in
CommandAssertUtils.

Change-Id: I8b737545c0519cc2034bb6eec36afcc38da26059
Bug-Url: https://bugzilla.redhat.com/863097
Signed-off-by: Vered Volansky <vvolansk@redhat.com>