core: use instance type in addVm + permissions
Added new ActionGroup CREATE_INSTANCE that allow user to create only
instance, which is more 'weak' than create vm because of the smaller set
of parameters the user can set to an instance compared to vm.

Added new roles to be used: InstanceCreator and UserInstanceManager
which are compatible with vm roles to allow instance manipulation by
users.

AddVmCommand: handle creation of instance (vm from instance type and
image type)

Change-Id: If9058907d1021b09981671e905defb91ed645d65
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>