engine: Fix SHD service not displaying issue (#885592)
To get the service info, the UI will not pass the volume name, in that case
engine will fetch the volume name in the database.

NFS volume name should be passed to get nfs service details, similarly
REPLICATE/DISTRIBUTED_REPLICATE volume name should be passed as an
argument to get the SHD details.

So to get volume name from database engine will do the following steps.
1. Engine will fetch NFS + REPLICATE/DISTRIBUTED_REPLICATE volume name.
2. If not found first the engine will fetch the nfs volume name and then fetch
   REPLICATE/DISTRIBUTED_REPLICATE volume name.
3. The VDS query will be called twice, one with nfs volume name
   and another with replicate volume name, finally combine the service details.

Change-Id: Id11725f44ab3fdd36f76fe569d7610a411518ee1
Bug-Url: https://bugzilla.redhat.com/885592
Signed-off-by: Dhandapani <dgopal@redhat.com>