core: ReconstructMasterDomain -connect/disconnect hosts simultaneously
When running ReconstructMasterDomain/RecoveryStoragePool commands - any
connect/disconnect operations can be done simultaneously by different
threads in order to improve the runtime.

There's no need to make use here of the runSynchronizeOperation() ,
it will require to add few classes, handle with parameter passing when
this flow is relevant to the Reconstruct flow, in many places in the
code we execute runnables using ThreadPoolUtils class so I prefer to
keep the implementation simple rather then complicate it and add classes that
will be never used again just to use CountDownLatch.

Change-Id: I3c8dac368d60eac85a92762dffc5dbb8062b85ef
Signed-off-by: Liron Aravot <laravot@redhat.com>