core: OvfDataUpdater - removal of update/remove OVF vdsm calls
This patch replaces calls to to following commands with a call to
incrementDbGeneration in order to singal OvfDataUpdater that the
given vm/template has been updated.

Generally, most commands are performed under lock to a specific
vm/template, but not all of them - as the row which contains the db
generation of a vm/template is accessed from many commands that perform
operation on it ,
the update statement to this row has been placed in the start of the db
update transaction
in order to avoid possible db deadlocks and errors.

http://wiki.ovirt.org/Feature/OvfAutoUpdater

Change-Id: Iedebceb9809dc0b11c0bbe8a2d4af63b0d848df1
Signed-off-by: Liron Aravot <laravot@redhat.com>