webadmin: Remove Load bar
Load Bar in the webadmin Hosts main tab is showing the number of vm
running on the host. This information is not related directly to the
load on the host in any way and it is just confusing.

The load bar was removed from view. Instead, a text column telling the
number of vm running on the host was added.

Change-Id: I1e9506c04975854e17508b6bebfe3c9044f21ca5
Signed-off-by: Ofri Masad <omasad@redhat.com>