engine: Forced removal of a host (#882807)
If all the hosts are down in a given cluster and the force remove action is true,
then the engine will remove the hosts entries in the db.

Also, if the removing host is the last host and the force remove action is true,
then clear the gluster volumes from the database.

Added new stored procedure to remove the volumes of a given cluster id.

Change-Id: I6939d3ee7eab38cb6b26abd98d7377be64cd9bdc
Bug-Url: https://bugzilla.redhat.com/882807
Signed-off-by: Dhandapani <dgopal@redhat.com>