engine: Extract SetNetworkStatus method
This method is used in a variety of places and is not necessarily
related to AttachNetworkToVdsGroupCommand. This is why it was moved to a
dedicated class that will hold such methods.

Some warnings were fixed on the new method, and the name was changed to
comply with Java naming conventions and with meaning.

Change-Id: Ia8f0e864c82b3977e1428e4854d85ca4c6e8a7ff
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>