restapi: remove usb check from round trip tests
Remove the check whether the usb's enabled state remains the same in
round trip mapping test of VmMapper and TemplateMapper.

This check is incorrect because:

1. in case the received usb setting is enabled but with no valid type,
it could be mapped to disabled (if it's a request to add new entity or if
the previous usb setting of the entity was disabled) and remains disabled
while mapping it back to rest api VM. thus checking that the enabled
state of the given usb setting remains the same in the round trip mapping
without checking that its type is valid is wrong.

2. the usb mapping was removed from the map methods of VmMapper and
TemplateMapper classes, as it's overridden by the caller anyway (because
the usb mapping requires information that cannot be access from inside of
the map methods). so it makes no sense to check the usb settings it in
this test that tests the map methods.

Change-Id: Ib7ce32dca6821a8b6eb8d69fe18413c87a6f0b3f
Signed-off-by: Arik Hadas <ahadas@redhat.com>