webadmin: Avoid duplicate server [Search] queries
While executing several queries in parallel, the UI gets stuck;
There is a hard limitation of ~6 connections
for http requests (for the same host name).

We are getting a lot of performance issues that are related to it.
The main reproducer to 'UI not responding' is
re-sending a 'slow' query and run out of connections.

For a large scale env, this can be a serious problem;

Currently we support only search queries because it's easy to identify
them; i.e., according to search string and sender, whereas other queries,
are pretty hard to compare, because 'equals' isn't implemented in
parameters classes.

In the patch:
When sending a second duplicate request from the same
place (model, queryType and parameters),
while the first request is waiting for response,
we will postpone it till the first is finished.
NOTE: we will postpone (save) only a single request, and avoid the others;
e.g., when pressing 30 time on a key in the GUI the first one will
be excuted and the others will pile up on the same place in the set.

Change-Id: I483a1c449f1eb2e9820464dbaf0a686f90923bb6
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>