engine: consolidate VM and Poll queries
Created a singe query 'GetAllVmsAndVmPoolsQuery', that wraps
the former 2 queries for vms and pools in the userportal.
These queries are called by the UP simultaneously, so
it is better to make a single query instead.

Reduced getPoolForUser user level since it asks for userId, and
no longer getting called from UP.

Change-Id: I59346bc9f1cf6d948c4401ddbd7bb365c67b5b05
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>