restapi: Vnic's network update when portMirroring is set behaves wrong
If port mirroring is set, the port mirroring network should be equal to
the vnic's network. Otherwise, an exception should be thrown.
There was a bug, the equlity was checked just if the port mirroring was
updated.
So, in case that the network was updated and the port mirroring wasn't-
an exception wasn't thrown although it should be.

Change-Id: Icbb346805f91bd3964550a19b18b3d0a0c40e12c
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>