core: Upgrade will update image statuses to OK.
As part of the upgrade process, engine cancel the tasks and
removes them.
part of this behaviour we also want to change the status of the disk
to OK, to avoid cases that the VM cannot be restarted.

The proposed solution changes all the LOCKED statuses of images to OK.

Change-Id: I2a6bd270ebd2125252a4b155f050db103969cb5e
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>