restapi: Activate/Deactivate vnic should use edit
As part of the Network Linking feature-
backend ActivateDeactivateVmNicCommand should become internal.
Plugging/Unplugging the nic will be done through the edit.

In order to keep backward compatibility in the rest, we keep the activate/
deactivate actions and deprecate them.
Their implementation is changed to use "edit" instead of
"ActivateDeactivateVmNicCommand"

Change-Id: Id6ac3ff0bf6c49aaea8dc589ca5da6495feb2d5b
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>