engine,webadmin: Networks sort should be done on engine side
Networks main tab should be sorted in the following way-
Primary sort- data center name
Secondary sort- network name

The sort was done on the webadmin side.
It caused a bug, cause if the search string included "sort by" it was
ignored (the webadmin sorted all the items ignoring the search string).

Change-Id: Ic9735606231e1947e76bd5550cbcc6566ac8a968
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>