core: fixed video device setting on import vm/template
This patch solves a bug in which the wrong video device was set for
vm/template that was exported in a release prior to 3.1 and imported
in release 3.1 or later.

The logic that selected the video device, in case it doesn't appear in
the ovf file of the vm/template, according to the number of
monitors of the vm/template was replaced. the logic now is to set it
according to the default display type of the vm/template, and only if
the default display type doesn't exist (should never happen though),
the video device would be decided according to the number of monitors.

The ovf file of vm/template which is exported in a release prior to 3.1
doesn't include video devices, so the video devices were determined
according to the number of monitors. if a vm had 1 monitor, cirrus
device was set for it - thus, desktop vms with one monitor or server
vms were given cirrus devices even when spice console was defined
for them. from now on they will have qxl device.

Change-Id: Id1bfaa3414507a982b9f22aa21fb3d05fe87e84d
Bug-Url: https://bugzilla.redhat.com/889499
Signed-off-by: Arik Hadas <ahadas@redhat.com>