core: create db fails on rhel 6.4 using postgres..
create db fails on rhel 6.4 using postgresql-8.4.13-1.el6_3.x86_64

This patch fix a failure on postgres 8.4 that fails to update a
parameter to a SP on postgres 8.4 while the same modification is OK for
postgres 9.x.

The solution uses a local SO variable to do the modifications and be
used later on on the INSERT statement.
This patch also modified the message data type to be text and not
varchar(4000) to match the audit_table definitions.

Change-Id: I4451f5bdf118efc02bb9100c59e0729283c4c2b5
Signed-off-by: Eli Mesika <emesika@redhat.com>
Bug-Url:https://bugzilla.redhat.com/show_bug.cgi?id=890982