core: fix run once starts with wrong display type
This patch fix the following bug: trying to run once a VM which has vnc
console set in its default settings, with spice console results in starting
qemu process with '-vga cirrus' instead of '-vga qxl'.

There was a special treatment for the case of run once with vnc console
a vm that has spice console in its default settings, but no treatment
for the scenario described above.

The fix is to add qxl video device in that case.

Change-Id: I635307847edf764b174230203ee355987b948ce4
Signed-off-by: Arik Hadas <ahadas@redhat.com>