packaging: Updated all utilities to use backup.sh instead of pg_dump
* Previously, the setup utilities used pg_dump command to backup the
database. The more appropriate way is to use the backup.sh utility,
that runs the dump of the database in a more correct and way. This
also has the added benifit in that changes in DB mechanisms or parameters
for dumping the data would be handled by the external script, and
changes to the setup code would not be required.

Bug-Url: https://bugzilla.redhat.com/866964
Change-Id: I011bc64d989cfba602977fd3f9522ea2c329aedc
Signed-off-by: Alex Lourie <alourie@redhat.com>