core: prevent db updates due to change in lastStartTime
This patch prevents unneccesary updates of VMs by VdsUpdateRunTimeInfo
monitoring in case the 'lastStartTime' attribute is the only attribute
of the VM that is changed (received with a different value from vdsm).

The 'lastStartTime' attribute is always received from vdsm with null
value, which is different from its value in the engine, thus
VdsUpdateRunTimeInfo thinks that the VM is changed and update it in
the DB (and that happens in every monitoring iteration).

The solution is to remove 'lastStartTime' attribute from the list
of attributes which are, when received from vdsm, compared to their
values in the engine in order to decide whether the VM is changed.

Change-Id: I8193defc1833e33bb3f8809d8416012c6dc27aea
Signed-off-by: Arik Hadas <ahadas@redhat.com>