core: rerun HA vm that fails during live snapshot
This patch fix a bug that happens when HA vm fails during live
snapshot operation, and stays down.

The problem was that when the monitoring detected that the vm is dowm,
it tried to rerun it, but the RunVm command failed because the vm was
during snapshot.

The solution is to save the VM status when starting to execute the
CreateAllSnapshots command, and at the end check if the VM is HA +
it's down because of an error + it wasn't down when the command
started - that would mean the VM went down unexpectedly during the
snapshot command, and thus will try to run it.

Change-Id: Id5daeffd1cee106019065c29fb8428a9b9f78a5d
Bug-Url: https://bugzilla.redhat.com/878041
Signed-off-by: Arik Hadas <ahadas@redhat.com>