userportal,webadmin: Allow e1000 nic type for Windows VM
Because of some licencing issues, the e1000 NIC was
not allowed to be selected for the windows VMs.

As this has been solved now, enabling it again.

This has already been fixed in 713361fc541874342b5540bff9f74b206dfc6a6f
but the 0c66da61a5d4d57299525737063bb57561134f5d have
deleted the DataProvider and have not ported this change to
AsyncDataProvider

Change-Id: I7f63cba7215e5b4615f949c4f26c084f349b0d73
Bug-Url: https://bugzilla.redhat.com/838470
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>