core: persist LUN with correct volume_group_id
When saving a LUN object to the db, we specify what is the needed volume
group. In case of an existing LUN - its volume group id would be updated to
the new given value. In case that the given LUN isn't in the DB, it'll
be saved. This patch ensures that the saved LUN would be saved with the
given volume_group_id.

NOTES:

The motivation for this change is that in cases in which we get the LUN
info from VDSM and the LUN is part of an existing VG, saving it with
unknown VG to the system will prevent removal of the LUN when it won't
be used anymore by the system and might cause confusion and inconsistency.

Change-Id: Ia4404b0d17b107208e5ec00c688db305ba0e65fe
Signed-off-by: Liron Aravot <laravot@redhat.com>