engine: Allow updating network if not used by hosts
The requirement for updating network only if not attached to
a cluster was mitigated and now we support updating a network
only if it is not being used (e.g. not attached to any host's nic):
As long as the network isn't defined on the hosts, there shouldn't
be an issue of modifying its configuration.

Change-Id: I75f6e21defef5e70e91013f9bd645e5d7d265ba8
Signed-off-by: Moti Asayag <masayag@redhat.com>