core:Can not upgrade from 3.1 to 3.2
This patch fix a bug that resulted from commit
6891171a906f708b8751cc583faffc94590e7948

In the above commit we had removed the external UUId package dependency
and replaced the implementation with our implementation keeping the same
function name.
But since the old original function was installed with user postgres and
we are trying to replace it with another user (engine), we got a message
that we have no privileges to drop the old function:
ERROR:  must be owner of function uuid_generate_v1

This fix runs a cleanup script that removes all old installed UUID
functions using the user postgres.

Change-Id: Iefbd051b30917827dd5dcf357cfcb5a6934ff589
Signed-off-by: Eli Mesika <emesika@redhat.com>
Bug-Url:https://bugzilla.redhat.com/show_bug.cgi?id=890324
Signed-off-by: Eli Mesika <emesika@redhat.com>